renamed structureverifier as fileverify; updated docs
