+++ dquality/reporter.py
@@ -60,7 +60,7 @@ import h5py
 from multiprocessing import Process, Queue
 from configobj import ConfigObj
 
-from structureverifier import layout
+from fileverify import layout
 from common.qualitychecks import Data, validate_mean_signal_intensity, validate_signal_intensity_standard_deviation, validate_voxel_based_SNR, validate_slice_based_SNR
 from common.utilities import get_data
